Add a point field to board item for item exchange
This makes sure at least itemId, which is just a string,
and its point can be exchanged between the board and the deck.

Note that we don't care item identity on client as long as it
maintains the rules of the game. Every data submitted from clients
will be validated on backend. So there is no worry of clients' hacks