Fix draggable on tablet and mobile.
There is a touch event problem when it is on mobile screen. Somehow
click event isn't fired when selecting a choice. It seems it relates
to z-index or parent/child nodes but it works on desktop. So, we fix
it by assuming both are the causes, and make 2 types of events
separeated - choice click and draggable.